+++ src/Opulence/Ioc/Bootstrappers/Factories/BootstrapperRegistryFactory.php
@@ -12,42 +12,21 @@ namespace Opulence\Ioc\Bootstrappers\Factories;
 
 use Opulence\Ioc\Bootstrappers\BootstrapperRegistry;
 use Opulence\Ioc\Bootstrappers\IBootstrapperRegistry;
-use Opulence\Ioc\Bootstrappers\IBootstrapperResolver;
-use Opulence\Ioc\Bootstrappers\LazyBootstrapper;
 
 /**
  * Defines the bootstrapper registry factory
  */
 class BootstrapperRegistryFactory implements IBootstrapperRegistryFactory
 {
-    /** @var IBootstrapperResolver The bootstrapper resolver */
-    protected $bootstrapperResolver = null;
-
-    /**
-     * @param IBootstrapperResolver $bootstrapperResolver The bootstrapper resolver
-     */
-    public function __construct(IBootstrapperResolver $bootstrapperResolver)
-    {
-        $this->bootstrapperResolver = $bootstrapperResolver;
-    }
-
     /**
      * @inheritdoc
      */
     public function createBootstrapperRegistry(array $bootstrapperClasses) : IBootstrapperRegistry
     {
-        $bootstrapperObjects = $this->bootstrapperResolver->resolveMany($bootstrapperClasses);
         $bootstrapperRegistry = new BootstrapperRegistry();
 
-        foreach ($bootstrapperObjects as $bootstrapperObject) {
-            if ($bootstrapperObject instanceof LazyBootstrapper) {
-                $bootstrapperRegistry->registerLazyBootstrapper(
-                    $bootstrapperObject->getBindings(),
-                    get_class($bootstrapperObject)
-                );
-            } else {
-                $bootstrapperRegistry->registerEagerBootstrapper(get_class($bootstrapperObject));
-            }
+        foreach ($bootstrapperClasses as $bootstrapperClass) {
+            $bootstrapperRegistry->registerBootstrapper(new $bootstrapperClass);
         }
 
         return $bootstrapperRegistry;