+++ src/Opulence/Ioc/Tests/Bootstrappers/Factories/BootstrapperRegistryFactoryTest.php
@@ -11,7 +11,6 @@
 namespace Opulence\Ioc\Tests\Bootstrappers\Factories;
 
 use Opulence\Ioc\Bootstrappers\Factories\BootstrapperRegistryFactory;
-use Opulence\Ioc\Bootstrappers\IBootstrapperResolver;
 use Opulence\Ioc\Tests\Bootstrappers\Mocks\Bootstrapper;
 use Opulence\Ioc\Tests\Bootstrappers\Mocks\BootstrapperWithEverything;
 use Opulence\Ioc\Tests\Bootstrappers\Mocks\LazyFooInterface;
@@ -23,16 +22,13 @@ class BootstrapperRegistryFactoryTest extends \PHPUnit\Framework\TestCase
 {
     /** @var BootstrapperRegistryFactory The factory to use in tests */
     private $factory = null;
-    /** @var IBootstrapperResolver|\PHPUnit_Framework_MockObject_MockObject The bootstrapper resolver */
-    private $resolver = null;
 
     /**
      * Sets up the tests
      */
     public function setUp() : void
     {
-        $this->resolver = $this->createMock(IBootstrapperResolver::class);
-        $this->factory = new BootstrapperRegistryFactory($this->resolver);
+        $this->factory = new BootstrapperRegistryFactory();
     }
 
     /**
@@ -41,11 +37,6 @@ class BootstrapperRegistryFactoryTest extends \PHPUnit\Framework\TestCase
     public function testCreatingRegistryResolvesBootstrappers() : void
     {
         $bootstrapperClasses = [Bootstrapper::class, BootstrapperWithEverything::class];
-        $bootstrapperObjects = [new Bootstrapper(), new BootstrapperWithEverything()];
-        $this->resolver->expects($this->any())
-            ->method('resolveMany')
-            ->with($bootstrapperClasses)
-            ->willReturn($bootstrapperObjects);
         $bootstrapperRegistry = $this->factory->createBootstrapperRegistry($bootstrapperClasses);
         $this->assertEquals([Bootstrapper::class], $bootstrapperRegistry->getEagerBootstrappers());
         $this->assertEquals(