+++ src/Opulence/Ioc/Tests/Bootstrappers/Factories/CachedBootstrapperRegistryFactoryTest.php
@@ -14,7 +14,6 @@ use Opulence\Ioc\Bootstrappers\BootstrapperRegistry;
 use Opulence\Ioc\Bootstrappers\Caching\ICache;
 use Opulence\Ioc\Bootstrappers\Factories\CachedBootstrapperRegistryFactory;
 use Opulence\Ioc\Bootstrappers\IBootstrapperRegistry;
-use Opulence\Ioc\Bootstrappers\IBootstrapperResolver;
 use Opulence\Ioc\Tests\Bootstrappers\Mocks\Bootstrapper;
 use Opulence\Ioc\Tests\Bootstrappers\Mocks\BootstrapperWithEverything;
 use Opulence\Ioc\Tests\Bootstrappers\Mocks\LazyFooInterface;
@@ -26,8 +25,6 @@ class CachedBootstrapperRegistryFactoryTest extends \PHPUnit\Framework\TestCase
 {
     /** @var CachedBootstrapperRegistryFactory The factory to use in tests */
     private $factory = null;
-    /** @var IBootstrapperResolver|\PHPUnit_Framework_MockObject_MockObject The bootstrapper resolver */
-    private $resolver = null;
     /** @var ICache|\PHPUnit_Framework_MockObject_MockObject The bootstrapper registry cache */
     private $cache = null;
 
@@ -36,9 +33,8 @@ class CachedBootstrapperRegistryFactoryTest extends \PHPUnit\Framework\TestCase
      */
     public function setUp() : void
     {
-        $this->resolver = $this->createMock(IBootstrapperResolver::class);
         $this->cache = $this->createMock(ICache::class);
-        $this->factory = new CachedBootstrapperRegistryFactory($this->resolver, $this->cache);
+        $this->factory = new CachedBootstrapperRegistryFactory($this->cache);
     }
 
     /**
@@ -59,17 +55,12 @@ class CachedBootstrapperRegistryFactoryTest extends \PHPUnit\Framework\TestCase
     public function testCacheMissManuallyCreatesRegistry() : void
     {
         $bootstrapperClasses = [Bootstrapper::class, BootstrapperWithEverything::class];
-        $bootstrapperObjects = [new Bootstrapper(), new BootstrapperWithEverything()];
         $lazyBootstrapperBindings = [
             LazyFooInterface::class => [
                 'bootstrapper' => BootstrapperWithEverything::class,
                 'target' => null
             ]
         ];
-        $this->resolver->expects($this->any())
-            ->method('resolveMany')
-            ->with($bootstrapperClasses)
-            ->willReturn($bootstrapperObjects);
         $this->cache->expects($this->any())
             ->method('get')
             ->willReturn(null);